refactor(agent): extract memory + goals tool handler from executor
Move the memory-capture and goal-management tool handlers out of
AgentToolExecutorService into a new AgentMemoryGoalsToolHandler,
following the AgentCompletionCardBuilderService (#1583) and
AgentThreadEventRecorderService extraction pattern.

Moved verbatim (pure move, no behavior changes):
- captureMemory, createGoal, checkGoalProgress, updateGoal
  (made public; were private async on the executor)
- normalizeAgentMemoryContentType, normalizeAgentMemoryKind,
  normalizeAgentMemoryScope (kept private; only used by captureMemory)

The executor now injects AgentMemoryGoalsToolHandler as a required
constructor dependency and delegates the four CAPTURE_MEMORY /
CREATE_GOAL / CHECK_GOAL_PROGRESS / UPDATE_GOAL dispatch cases to it.
This removes the executor's two now-dead @Optional() dependencies,
agentMemoryCaptureService and agentGoalsService, which are no longer
referenced anywhere in the executor.

createBrand and its checkOnboardingStatus/normalizeHandle helpers are
a different tool family and were intentionally left untouched.

Executor: 9307 -> 8986 lines.

Behavior preserved: all guard clauses, error strings, and result
payload shapes are unchanged from the original private methods.

Verification: local tests/typecheck/build are intentionally not run
per repo policy (MacBook verification constraint) — PR CI is the
source of truth. Both AgentToolExecutorService construction sites in
its spec were updated to construct/pass AgentMemoryGoalsToolHandler
instead of the two removed services directly; a new spec file covers
all four public methods' present/absent/success/failure paths.

Part of #519.
