+++ apps/server/api/src/services/agent-orchestrator/agent-orchestrator.module.ts
@@ -35,6 +35,7 @@ import { AgentOrchestratorController } from '@api/services/agent-orchestrator/ag
 import { AgentOrchestratorService } from '@api/services/agent-orchestrator/agent-orchestrator.service';
 import { AgentStreamPublisherModule } from '@api/services/agent-orchestrator/agent-stream-publisher.module';
 import { AgentToolsController } from '@api/services/agent-orchestrator/agent-tools.controller';
+import { AgentMemoryGoalsToolHandler } from '@api/services/agent-orchestrator/tools/agent-memory-goals-tool-handler.service';
 import { AgentRouteRewriteService } from '@api/services/agent-orchestrator/tools/agent-route-rewrite.service';
 import { AgentToolExecutorService } from '@api/services/agent-orchestrator/tools/agent-tool-executor.service';
 import { AgentSpawnModule } from '@api/services/agent-spawn/agent-spawn.module';
@@ -96,6 +97,7 @@ import { forwardRef, Module } from '@nestjs/common';
   ],
   providers: [
     AgentCompletionCardBuilderService,
+    AgentMemoryGoalsToolHandler,
     AgentOrchestratorService,
     AgentRouteRewriteService,
     AgentToolExecutorService,
